Move names related utilities to `names.h`
