+++ src/vctrs.h
@@ -244,6 +244,7 @@ bool vec_is_unspecified(SEXP x);
 // Vector methods ------------------------------------------------
 
 #include "arg.h"
+#include "names.h"
 
 enum vctrs_proxy_kind {
   vctrs_proxy_default,
@@ -401,22 +402,6 @@ enum vctrs_dbl_class {
 };
 
 enum vctrs_dbl_class dbl_classify(double x);
-
-// Names --------------------------------------------------------
-
-enum name_repair_arg {
-  name_repair_none,
-  name_repair_minimal,
-  name_repair_unique,
-  name_repair_universal,
-  name_repair_check_unique
-};
-
-const char* name_repair_arg_as_c_string(enum name_repair_arg arg);
-enum name_repair_arg validate_name_repair(SEXP arg);
-SEXP vec_as_names(SEXP names, enum name_repair_arg type, bool quiet);
-bool is_unique_names(SEXP names);
-SEXP vec_as_unique_names(SEXP names, bool quiet);
 
 // Character translation ----------------------------------------
